+++ Cargo-minimal.lock
@@ -55,6 +55,7 @@ dependencies = [
  "bech32",
  "bincode",
  "bitcoin-consensus-encoding",
+ "bitcoin-crypto",
  "bitcoin-internals",
  "bitcoin-io",
  "bitcoin-network-kind",
@@ -89,6 +90,11 @@ dependencies = [
 [[package]]
 name = "bitcoin-crypto"
 version = "0.0.0"
+dependencies = [
+ "arbitrary",
+ "bitcoin-internals",
+ "serde",
+]
 
 [[package]]
 name = "bitcoin-fuzz"
