+++ Cargo-recent.lock
@@ -54,6 +54,7 @@ dependencies = [
  "bech32",
  "bincode",
  "bitcoin-consensus-encoding",
+ "bitcoin-crypto",
  "bitcoin-internals",
  "bitcoin-io",
  "bitcoin-network-kind",
@@ -88,6 +89,11 @@ dependencies = [
 [[package]]
 name = "bitcoin-crypto"
 version = "0.0.0"
+dependencies = [
+ "arbitrary",
+ "bitcoin-internals",
+ "serde",
+]
 
 [[package]]
 name = "bitcoin-fuzz"
