+++ crypto/src/lib.rs
@@ -14,3 +14,6 @@ extern crate alloc;
 
 #[cfg(feature = "std")]
 extern crate std;
+
+#[cfg(feature = "alloc")]
+pub mod sighash;
